Move the implementation of transform_r_to_u_cell from MappingQ1 to MappingQGeneric.
